+++ packages/reactabular-resolve/src/_index.js
@@ -1,8 +1,8 @@
-function index({ rowData, rowIndex }) {
-  return {
+function index({ rowIndex }) {
+  return rowData => ({
     ...rowData,
     _index: rowIndex
-  };
+  });
 }
 
 export default index;