+++ packages/reactabular-resolve/src/nested.js
@@ -1,25 +1,27 @@
 import get from 'lodash/get';
 import has from 'lodash/has';
 
-function nested({ rowData, column }) {
-  const { property } = column;
+function nested({ column }) {
+  return (rowData) => {
+    const { property } = column;
 
-  if (!property) {
-    return {};
-  }
+    if (!property) {
+      return {};
+    }
 
-  if (!has(rowData, property)) {
-    console.warn( // eslint-disable-line no-console
-      `resolve.nested - Failed to find "${property}" property from`,
-      rowData
-    );
+    if (!has(rowData, property)) {
+      console.warn( // eslint-disable-line no-console
+        `resolve.nested - Failed to find "${property}" property from`,
+        rowData
+      );
 
-    return {};
-  }
+      return {};
+    }
 
-  return {
-    ...rowData,
-    [property]: get(rowData, property)
+    return {
+      ...rowData,
+      [property]: get(rowData, property)
+    };
   };
 }